Move slice canonicalizations into Slice.cpp
